Controllers

+++ src/main/java/com/bookstore/bookstore/repository/BookRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.List;
 import java.util.Set;
 
-public interface BookRepository extends MongoRepository<Book, Long> {
+public interface BookRepository extends MongoRepository<Book, Integer> {
 
     Set<Book> findByAuthor(String author);
